+++ R/filesstrings.R
@@ -2,6 +2,7 @@
 #' @importFrom magrittr '%>%' '%<>%' '%T>%'
 #' @useDynLib filesstrings, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
+#' @importFrom strex match_arg
 NULL
 
 ## quiets concerns of R CMD check re: the .'s that appear in pipelines
@@ -12,17 +13,6 @@ if (getRversion() >= "2.15.1") {
 .onUnload <- function (libpath) {
   library.dynam.unload("filesstrings", libpath)
 }
-
-#' Defunct functions
-#'
-#' These functions have been made defunct, mostly because the naming style of
-#' the package has been changed. Some have been removed because they were not
-#' well-done.
-#'
-#' @param ... Defunct function arguments.
-#'
-#' @name filesstrings-defunct
-NULL
 
 #' `filesstrings`: handy file and string manipulation
 #'